perf(overlay): the application-wide dispatch door bounded a COUNT of whole pictures where each one is 1.5x1.5 viewports at four bytes a texel, so the same ceiling of four was 71.2 MiB of pipe at 1920x1080 and 170.1 MiB at 3440x1440
`MAX_OVERLAY_PICTURES_OUTSTANDING = 4` bounded the number of whole-picture
overlay rasters the pipe could hold, and what four of them cost is a property
of the display: `OVERLAY_OVERSAMPLE_PERCENTS[0] = 150` makes every picture
1.5 x 1.5 viewports, so the door's real ceiling grew with the square of the
side and was never chosen at any canvas but the one it was reasoned at.

It is now `MAX_OVERLAY_PICTURE_BYTES_OUTSTANDING`, 50,331,648 B — twelve
4 MiB band drains, where eight is the 133 ms rasterize a pipe must cover to
keep the band queue from running dry. At 1920x1080 that is 3 pictures and
53.4 MiB against 71.2, at 2878x1651 2 and 81.6 against 163.1, at 3440x1440 2
and 85.0 against 170.1.

The count's own note argued against this — "a byte budget would re-derive the
plan here" — and nothing is re-derived: the charge is `OverlayTexturePlan::
bytes()` of the plan the door is already holding when it decides, carried onto
the cache so the in-flight half reads the same figure back, and an arrived
picture is priced off its own width and height.

The door admits while the pipe is BELOW the line rather than while the next
picture fits under it. "Fits" is the door that can refuse everything: a
picture larger than the ceiling never fits at any occupancy, and the pipe
would idle for ever on a canvas nobody had priced.

`fit::NeedTerms::upload_pending_host` stays at a whole batch, and the door
does not make it re-priceable: `App::arrived_overlay_asks` is a second
producer into the same dispatch which refuses only on this cache's own hold
and on `RendersInFlight::admits`, and never consults the aggregate allowance
at all, so one arrival round can still put a whole batch in flight outside the
ceiling. Its doc now says that instead of the claim it carried.

The pane's charge also stops merging the two states: a layer can hold an
arrived picture while a fresh one is being rasterized for it — which is what
the arrival door creates — and both are resident.

+++ squallar-egui/src/ui_map.rs
@@ -195,9 +195,10 @@ impl super::Gui {
         // One read for every pane this frame draws: the figure is the
         // device's, not the pane's.
         let overlay_render_limit = self.concurrent_renders;
-        // **What this frame may still ask the rasterizer for**, for the whole
-        // application rather than for one pane and layer. Opened here, before
-        // any pane is drawn, because the quantity it bounds is the batch —
+        // **What this frame may still ask the rasterizer for**, in host bytes
+        // and for the whole application rather than for one pane and layer.
+        // Opened here, before any pane is drawn, because the quantity it
+        // bounds is the batch —
         // every shown layer of every pane re-rasterizes together on a move —
         // and spent by `render_pane_map_content`'s door. See
         // `Gui::overlay_dispatch_budget`.
@@ -2265,7 +2266,7 @@ struct FloorStripCtx<'a> {
     /// the strip draws the same pane's layers and asks the same door, so a
     /// second allowance here would let one frame dispatch twice the batch.
     /// See [`pane_render::PaneRenderCtx::overlay_dispatch_budget`].
-    overlay_dispatch_budget: &'a std::cell::Cell<usize>,
+    overlay_dispatch_budget: &'a std::cell::Cell<u64>,
     pane: &'a mut crate::pane::PaneState,
     /// The viewport the strip is drawn through, **owned**.
     map_memory: walkers::MapMemory,
